requirements.txt
torch 
torchvision 
torchaudio --index-url https://download.pytorch.org/whl/cu118
Pillow==9.5.0
diffusers 
transformers 
accelerate 
scipy 
safetensors
cuda-python